feat: Switch UiAutomator2 server to W3C protocol (#294)
BREAKING CHANGE: session capabilities in W3C format are the only accepted
BREAKING CHANGE: API responses are formatted according to W3C format
element keys are currently accepted and returned for both JSONWP and W3C formats

+++ app/src/main/java/io/appium/uiautomator2/core/EventRegister.java
@@ -73,7 +73,7 @@ public abstract class EventRegister {
 
     public static Boolean runAndRegisterScrollEvents(ReturningRunnable<Boolean> runnable) {
         Session session = AppiumUIA2Driver.getInstance().getSessionOrThrow();
-        int timeout = (int) session.getCapability(EVENT_COOLDOWN_CAP, EVENT_COOLDOWN_MS);
+        int timeout = session.getCapability(EVENT_COOLDOWN_CAP, EVENT_COOLDOWN_MS);
 
         return runAndRegisterScrollEvents(runnable, timeout);
     }
